// src/app/api/model/updateCategoryRequest.ts
/**
 * FinanzApp API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Cuerpo de PATCH /categories/{category_id}
 */
export interface UpdateCategoryRequest {
  name?: string | null;
  parent_id?: string | null;
  color?: string | null;
  icon?: string | null;
  is_active?: boolean | null;
}